+++ packages/design-system-react-native/src/components/Text/Text.stories.tsx
@@ -124,7 +124,7 @@ export const FontWeightStory: Story = {
 export const FontFamilyStory: Story = {
   render: () => (
     <View>
-      <Text fontFamily={FontFamily.Default}>Default (Centra No 1)</Text>
+      <Text fontFamily={FontFamily.Default}>Default (Geist)</Text>
       <Text fontFamily={FontFamily.Accent}>Accent (MM Sans)</Text>
       <Text fontFamily={FontFamily.Hero}>Hero (MM Poly)</Text>
     </View>